+++ lib/server/generate.js
@@ -8,6 +8,8 @@
 async function execute() {
   const extractTranslations = require('../write-translations.js');
 
+  const metadataUtils = require('./metadataUtils');
+
   const CWD = process.cwd();
   const cssnano = require('cssnano');
   const fs = require('fs-extra');
@@ -147,8 +149,9 @@ async function execute() {
       return;
     }
 
-    let rawContent = readMetadata.extractMetadata(fs.readFileSync(file, 'utf8'))
-      .rawContent;
+    let rawContent = metadataUtils.extractMetadata(
+      fs.readFileSync(file, 'utf8')
+    ).rawContent;
 
     const language = metadata.language;
 
@@ -261,7 +264,7 @@ async function execute() {
         .replace('-', '/')
         .replace('-', '/')
         .replace(/\.md$/, '.html');
-      const result = readMetadata.extractMetadata(
+      const result = metadataUtils.extractMetadata(
         fs.readFileSync(normalizedFile, {encoding: 'utf8'})
       );
       const rawContent = result.rawContent;
