+++ lib/version.js
@@ -15,6 +15,7 @@ const chalk = require('chalk');
 const readMetadata = require('./server/readMetadata.js');
 const utils = require('./server/utils.js');
 const versionFallback = require('./server/versionFallback.js');
+const metadataUtils = require('./server/metadataUtils.js');
 const env = require('./server/env.js');
 
 const CWD = process.cwd();
@@ -85,7 +86,7 @@ files.forEach(file => {
     return;
   }
 
-  const res = readMetadata.extractMetadata(fs.readFileSync(file, 'utf8'));
+  const res = metadataUtils.extractMetadata(fs.readFileSync(file, 'utf8'));
   let metadata = res.metadata;
   // Don't version any docs without any metadata whatsoever.
   if (Object.keys(metadata).length === 0) {
